Store Traits in cache at the end instead of the beginning
If we cache the values at the beginning and something goes wrong while sending attributes to the Braze SDK (and historically it has) or the current user is null (it can happen the first time),
we will end up with a corrupted cache. Because those Traits will not have been sent to Braze, but since they are cached they won't ever be sent again.

By caching the values at the end of the function we make sure they are only cached if nothing unexpected breaks.